src/main/java/com/atomicleopard/webelemental/matchers/HasTextMatcher.java
/*
 *  Copyright (c) 2011 Nicholas Okunew
 *  All rights reserved.
 *  
 *  This file is part of the com.atomicleopard.webelemental library
 *  
 *  The com.atomicleopard.webelemental library is free software: you 
 *  can redistribute it and/or modify it under the terms of the GNU
 *  Lesser General Public License as published by the Free Software Foundation, 
 *  either version 3 of the License, or (at your option) any later version.
 *  
 *  The com.atomicleopard.webelemental library is distributed in the hope
 *  that it will be useful, but WITHOUT ANY WARRANTY; without even
 *  the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 *  See the GNU Lesser General Public License for more details.
 *  
 *  You should have received a copy of the GNU Lesser General Public License
 *  along with the com.atomicleopard.webelemental library.  If not, see 
 *  http://www.gnu.org/licenses/lgpl-3.0.html.
 */
package com.atomicleopard.webelemental.matchers;

import org.hamcrest.BaseMatcher;
import org.hamcrest.Description;
import org.hamcrest.Matcher;

import com.atomicleopard.webelemental.Element;

public class HasTextMatcher extends BaseMatcher<Element> {

	private final Matcher<String> matcher;

	public HasTextMatcher(Matcher<String> matcher) {
		this.matcher = matcher;
	}

	public boolean matches(Object o) {
		return matcher.matches(((Element) o).text());
	}

	public void describeTo(Description description) {
		description.appendText("has text ");
		matcher.describeTo(description);
	}

}
